requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
boto3==1.12.31
botocore==1.15.31
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dictor==0.1.5
docutils==0.15
Flask==1.1.1
Flask-DotEnv==0.1.2
Flask-Inputs==0.3.0
flask-request-validator==2.1.2
future==0.18.2
googlemaps==4.2.0
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==3.0.0a1
jmespath==0.9.5
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.4.1
python-dateutil==2.8.1
python-dotenv==0.12.0
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.0
WTForms==2.2.1
zipp==3.1.0
